Use a new environment variable for CORS permissions
I changed the environment variable URL into ALLOW_ORIGIN_URL, and I use it as the value
for Access-Control-Allow-Origin, so I can change the environment variable value
on a deployment.yaml file instead of changing an image.  Previously I had hard-coded
the value for the Access-Control-Allow-Origin header on the server.ts files.