Adjust the container resources for cpu and memory.
To deploy the app to the cloud for the first time, I had to match the
cpu and memory requests and limits on the deployment .yaml files.
I chose to use t2.medium ec2 instances for the worker nodes, and
to fit the pods on the nodes, I reduced the cpu requests and limits.