Adjust deployment, docker-compose-build, and .travis files
Change files so Travis can build and deploy to AWS the v1.1 version of udagram.
I added a README.md, and changed any reference to the docker images to use
the v1.1 images for backend-user and backend-feed.
I also wrote the .travis.yml file to build the images and deploy to an AWS cluster.